+++ doc/changes.rst
@@ -55,6 +55,9 @@ New in 0.43
     documented with :attr:`pipeline.Pipeline.variables`.
   - :class:`pipeline.RawSource`: the ``rename_channels`` parameter now renames
     channels in the montage/adjacency instead of renaming channels in the data.
+  - :class:`pipeline.RawMaxwell` can compensate for head movement in recordings
+    with continuous HPI through the ``head_pos`` parameter. This requires
+    ``mne > 1.12.1``.
 
 * Boosting: